debugged stock_extractor, time_series, stock_prediction, added helper
