--- Stock Analysis & Prediction/Social Media/data_extractor.py
@@ -1,49 +1,0 @@
-import praw  # Reddit API Wrapper
-import tweepy  # Twitter API Wrapper
-import requests
-import pandas as pd
-
-class SocialMediaDataExtractor:
-    def __init__(self, reddit_credentials, twitter_credentials):
-        # Initialize Reddit API client
-        self.reddit = praw.Reddit(
-            client_id=reddit_credentials['client_id'],
-            client_secret=reddit_credentials['client_secret'],
-            user_agent=reddit_credentials['user_agent']
-        )
-        
-        # Initialize Twitter API client
-        auth = tweepy.OAuthHandler(twitter_credentials['api_key'], twitter_credentials['api_secret_key'])
-        auth.set_access_token(twitter_credentials['access_token'], twitter_credentials['access_token_secret'])
-        self.api = tweepy.API(auth)
-    
-    def fetch_reddit_data(self, subreddit_name, num_posts):
-        subreddit = self.reddit.subreddit(subreddit_name)
-        posts = []
-        for post in subreddit.hot(limit=num_posts):
-            posts.append([post.title, post.score, post.id, post.subreddit, post.url, post.num_comments, post.selftext, post.created])
-        posts = pd.DataFrame(posts, columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])
-        return posts
-
-    def fetch_twitter_data(self, hashtag, num_tweets):
-        tweets = []
-        for tweet in tweepy.Cursor(self.api.search_tweets, q=hashtag, lang="en", tweet_mode='extended').items(num_tweets):
-            tweets.append([tweet.created_at, tweet.id, tweet.full_text])
-        tweets = pd.DataFrame(tweets, columns=['created_at', 'tweet_id', 'text'])
-        return tweets
-    
-    def fetch_facebook_data(self, page_id, num_posts):
-        # Logic to fetch data from Facebook
-        pass
-    
-    def store_data(self, data, filename):
-        data.to_csv(filename, index=False)
-    
-    def analyze_data(self, data):
-        # Logic to analyze fetched data
-        pass
-    
-    def visualize_data(self, analysis_results):
-        # Logic to visualize analysis results
-        pass
-
